test(picker-internal): migrate basic
