+++ core/src/components/picker-internal/test/basic/picker-internal.e2e.ts-snapshots/picker-internal-inline-diff-ios-rtl-Mobile-Chrome-linux.png
