+++ core/src/components/picker-internal/test/basic/picker-internal.e2e.ts-snapshots/picker-internal-modal-diff-ios-ltr-Mobile-Chrome-linux.png
